backend/app/schemas/user.py
from datetime import datetime

from pydantic import BaseModel, EmailStr


class UserBase(BaseModel):
    email: EmailStr
    full_name: str | None = None


class UserCreate(UserBase):
    pass


class UserRead(UserBase):
    id: int
    is_email_verified: bool
    created_at: datetime

    model_config = {"from_attributes": True}
