backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anyalert-backend"
version = "0.1.0"
description = "FastAPI backend for AnyAlert"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "email-validator>=2.2.0",
  "sqlalchemy[asyncio]>=2.0.32",
  "asyncpg>=0.29.0",
  "alembic>=1.13.2",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt==4.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.8",
  "httpx>=0.27.0",
  "ruff>=0.6.1"
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.setuptools.packages.find]
include = ["app*"]
